Keep const.py to constants and move the name helper to the rt helper
Review on #174: const.py holds constants, not functions. safe_file_part
moves next to the vehicle positions writer, the other caller of it, and
gtfs_helper takes it from there through the import it already has, so
nothing new points the other way.
